+++ handlers.go
@@ -16,6 +16,7 @@ func Index(w http.ResponseWriter, r *http.Request) {
 }
 
 func PostIndex(w http.ResponseWriter, r *http.Request) {
+	fmt.Println("INDEXXX")
 	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
 	w.WriteHeader(http.StatusOK)
 	if err := json.NewEncoder(w).Encode(post.FindAll()); err != nil {
