--- todo.go
@@ -1,12 +1,0 @@
-package main
-
-import "time"
-
-type Todo struct {
-	Id        int       `json:"id"`
-	Name      string    `json:"name"`
-	Completed bool      `json:"completed"`
-	Due       time.Time `json:"due"`
-}
-
-type Todos []Todo
